requirements.txt
aiogram~=3.6.0
python-dotenv~=1.0.1
requests~=2.31.0
aiohttp~=3.9.3
aiosqlite~=0.20.0
SQLAlchemy~=2.0.28
pydantic~=2.6.3
pydantic-settings~=2.2.1
alembic~=1.13.1
APScheduler~=3.10.4